Ability to add price for each country

+++ database/migrations/2017_06_10_062632_add_foreign_keys_to_prices_table.php
@@ -10,10 +10,18 @@ class AddForeignKeysToPricesTable extends Migration
      *
      * @return void
      */
+
+ 
+
     public function up()
     {
         Schema::table('prices', function (Blueprint $table) {
-            $table->foreign('product_id')->references('id')->on('products')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('prices');
+            if(!array_key_exists("prices_product_id_foreign", $indexesFound)){
+                $table->foreign('product_id')->references('id')->on('products')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+
+            }
         });
     }
